requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
en-core-web-sm==3.0.0
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.17
idna==2.10
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
praw==7.3.0
prawcore==2.2.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.2
ptyprocess==0.7.0
pyarrow==4.0.1
pycparser==2.20
pydantic==1.7.4
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
Send2Trash==1.5.0
six==1.16.0
smart-open==3.0.0
smmap==4.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
streamlit==0.82.0
terminado==0.10.0
testpath==0.5.0
thinc==8.0.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==2.1
update-checker==0.18.0
urllib3==1.26.5
validators==0.18.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
en-core-web-sm==3.0.0
entrypoints==0.3
fastapi==0.65.2
gitdb==4.0.7
GitPython==3.1.17
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
praw==7.3.0
prawcore==2.2.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.2
ptyprocess==0.7.0
pyarrow==4.0.1
pycparser==2.20
pydantic==1.7.4
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
Send2Trash==1.5.0
six==1.16.0
smart-open==3.0.0
smmap==4.0.0
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.4.18
srsly==2.4.1
sse-starlette==0.6.2
starlette==0.14.2
streamlit==0.82.0
terminado==0.10.0
testpath==0.5.0
thinc==8.0.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==2.1
update-checker==0.18.0
urllib3==1.26.5
validators==0.18.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
